Moved freetype text & embed into separate files
