+++ client/app/scripts/liveblog-edit/directives/freetype-embed.js
@@ -1,0 +1,13 @@
+freetypeEmbed.$inject = ['$compile'];
+
+export default function freetypeEmbed($compile) {
+    return {
+        restrict: 'E',
+        template: '<textarea ng-model="embed" rows="8"></textarea>',
+        controller: function() {
+        },
+        scope: {
+            embed: '='
+        }
+    };
+}
